break the huge editor.js into smaller modules
