+++ src/js/edit/index.js
@@ -1,0 +1,10 @@
+import './util.js';
+import './simplehtml.js';
+import './comment.js';
+import './state.js';
+import './helpers.js';
+import './realtime.js';
+import './drafts.js';
+import './ui.js';
+import './tabs.js';
+import './actions.js';
